+++ src/main/java/it/eng/idsa/dataapp/configuration/ECCProperties.java
@@ -7,8 +7,10 @@ import org.springframework.stereotype.Component;
 @ConfigurationProperties("application.ecc")
 public class ECCProperties {
 	private String protocol;
+	private String RESTprotocol;
 	private String host;
 	private int port;
+	private int RESTport;
 	private String mixContext;
 	private String formContext;
 	private String headerContext;
@@ -18,6 +20,12 @@ public class ECCProperties {
 	}
 	public void setProtocol(String protocol) {
 		this.protocol = protocol;
+	}
+	public String getRESTprotocol() {
+		return RESTprotocol;
+	}
+	public void setRESTprotocol(String rESTprotocol) {
+		RESTprotocol = rESTprotocol;
 	}
 	public String getHost() {
 		return host;
@@ -30,6 +38,12 @@ public class ECCProperties {
 	}
 	public void setPort(int port) {
 		this.port = port;
+	}
+	public int getRESTport() {
+		return RESTport;
+	}
+	public void setRESTport(int RESTport) {
+		this.RESTport = RESTport;
 	}
 	public String getMixContext() {
 		return mixContext;