+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.8-SNAPSHOT] - 2021-07-09
+
+### Added
+ - DescriptionRequestMessage handling
+ - Rejection message for ArtifactRequestMessage if some custom criteria are not met (only for demonstration purposes)
+
+### Changed
+
 ## [0.0.7-SNAPSHOT] - 2021-07-08
 
 ### Added
@@ -12,7 +20,7 @@ All notable changes to this project will be documented in this file.
 ## [0.0.6-SNAPSHOT] - 2021-07-02
 
 ### Added
- - fix for handling proxy payload - String nad JSON support
+ - fix for handling proxy payload - String and JSON support
 
 ### Changed